Sauce settings updated May 20th data updates

+++ fidowebautomation/src/ca/fido/test/tests/selfserve/FidoSS_Regression_TC62_ValidateCancelDataFlowMultilineAHWithNonWPlanCTNWithSingleRegularMDT.java
@@ -70,9 +70,9 @@ public class FidoSS_Regression_TC62_ValidateCancelDataFlowMultilineAHWithNonWPla
 		Map<String, Integer> countOfActiveAndCancelledAddData = fido_data_management_page.getAllExistingAddDataCountCancelledAndActive();
 		reporter.reportLogWithScreenshot("Manage Data page");
 		//Comparisions Before Cancel:
-		reporter.softAssert((countOfActiveAndCancelledAddDataOnMyPlan.get("cancelled")==countOfActiveAndCancelledAddData.get("cancelled")
-							&& countOfActiveAndCancelledAddDataOnMyPlan.get("active")==countOfActiveAndCancelledAddData.get("active")
-							&& countOfActiveAndCancelledAddDataOnMyPlan.get("nonMDT")==countOfActiveAndCancelledAddData.get("nonMDT"))
+		reporter.softAssert((countOfActiveAndCancelledAddDataOnMyPlan.get("cancelled").intValue()==countOfActiveAndCancelledAddData.get("cancelled").intValue()
+							&& countOfActiveAndCancelledAddDataOnMyPlan.get("active").intValue()==countOfActiveAndCancelledAddData.get("active").intValue()
+							&& countOfActiveAndCancelledAddDataOnMyPlan.get("nonMDT").intValue()==countOfActiveAndCancelledAddData.get("nonMDT").intValue())
 				, "The number of cancelled and active add on macth on my plans and manage data page", 
 				"The number of cancelled and active add on does not macth on my plans and manage data page");
 		
@@ -81,7 +81,8 @@ public class FidoSS_Regression_TC62_ValidateCancelDataFlowMultilineAHWithNonWPla
 			fido_data_management_page.scrollToMiddle();
 			reporter.reportLogWithScreenshot("Click on cancel MDT Link");
 			fido_data_management_page.clkCancelMDTLink();			
-			reporter.reportLogWithScreenshot("Click on Yes Remove Top Up");
+
+reporter.reportLogWithScreenshot("Click on Yes Remove Top Up");
 			fido_data_management_page.clkYesRemoveTopUpButton();
 			reporter.hardAssert(fido_data_management_page.isCancelSuccessdisplayed(),
 					"Cancel MDT success",
